apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
psutil==6.0.0
supabase==2.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.3
